requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.2
django-environ==0.12.0
django-quill-editor==0.1.42
django-sendgrid-v5==1.3.0   
django-widget-tweaks==1.5.0
gTTS==2.5.4
gunicorn==23.0.0
idna==3.10
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pypdf==6.1.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
reportlab==4.4.4
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
